Extract PATH env warning
